1641: separate out logic into util file
