+++ docker/dockerfiles/Dockerfile.server
@@ -49,6 +49,7 @@ COPY mkdocs.yml mkdocs.yml
 COPY monitor.py monitor.py
 COPY scripts scripts
 COPY setup.py setup.py
+COPY test_util.py test_util.py
 COPY test_cli.py test_cli.py
 COPY stress_test.py stress_test.py
 COPY tests tests
